# docs/swagger.yaml
basePath: /
definitions:
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  v1.ProjectConfigResponse:
    properties:
      keys:
        items:
          $ref: '#/definitions/v1.ProjectKeyResponse'
        type: array
      project:
        $ref: '#/definitions/v1.ProjectResponse'
      tables:
        items:
          $ref: '#/definitions/v1.ProjectTableResponse'
        type: array
    type: object
  v1.ProjectKeyResponse:
    properties:
      expires_at:
        type: string
      id:
        type: integer
      issuer:
        type: string
      key_id:
        type: string
      project_id:
        type: string
      status:
        type: string
    type: object
  v1.ProjectResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      project_id:
        type: string
      project_name:
        type: string
      school:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  v1.ProjectTableResponse:
    properties:
      id:
        type: integer
      notice:
        type: boolean
      project_id:
        type: string
      scopes:
        items:
          type: string
        type: array
      status:
        type: string
      table_id:
        type: string
      table_identity:
        type: string
      table_name:
        type: string
      table_token:
        type: string
      table_type:
        type: string
      view_id:
        type: string
    type: object
  v1.RegisterProjectKeyReq:
    properties:
      expires_at:
        type: string
      issuer:
        type: string
      key_id:
        type: string
      public_key:
        type: string
    required:
    - issuer
    - key_id
    - public_key
    type: object
  v1.RegisterProjectReq:
    properties:
      key:
        $ref: '#/definitions/v1.RegisterProjectKeyReq'
      project_id:
        type: string
      project_name:
        type: string
      school:
        type: string
      status:
        type: string
      tables:
        items:
          $ref: '#/definitions/v1.RegisterProjectTableReq'
        minItems: 1
        type: array
    required:
    - key
    - project_id
    - project_name
    - school
    - tables
    type: object
  v1.RegisterProjectTableReq:
    properties:
      notice:
        type: boolean
      scopes:
        items:
          type: string
        minItems: 1
        type: array
      table_id:
        type: string
      table_identity:
        type: string
      table_name:
        type: string
      table_token:
        type: string
      table_type:
        type: string
      view_id:
        type: string
    required:
    - scopes
    - table_id
    - table_identity
    - table_name
    - table_token
    - view_id
    type: object
...
  v1.TriggerNotificationReq:
    properties:
      project_id:
        description: 项目标识
        type: string
      table_identify:
        description: 反馈表格 Identify，反馈表的唯一标识
        type: string
    required:
    - project_id
    - table_identify
    type: object
...
  v1.UpdateProjectConfigReq:
    properties:
      key:
        $ref: '#/definitions/v1.RegisterProjectKeyReq'
      project_name:
        type: string
      school:
        type: string
      status:
        type: string
      tables:
        items:
          $ref: '#/definitions/v1.RegisterProjectTableReq'
        minItems: 1
        type: array
    required:
    - key
    - project_name
    - school
    - status
    - tables
    type: object
  v1.UpdateProjectReq:
    properties:
      project_name:
        type: string
      school:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: 木犀反馈系统 API
  version: "1.0"
paths:
...
  /api/v1/integrations/projects:
    get:
      parameters:
      - description: Bearer Admin JWT
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/v1.ProjectResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取项目列表
      tags:
      - Integration Admin
    post:
      consumes:
      - application/json
      description: 登记项目身份公钥、飞书反馈表和表级 scope。接口仅允许管理员访问，返回完整项目配置。
      parameters:
      - description: Bearer Admin JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: 项目配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.RegisterProjectReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.ProjectConfigResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      summary: 注册反馈项目
      tags:
      - Integration Admin
  /api/v1/integrations/projects/{project_id}:
    delete:
      parameters:
      - description: Bearer Admin JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: 项目 ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除项目
      tags:
      - Integration Admin
    get:
      parameters:
      - description: Bearer Admin JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: 项目 ID
        in: path
        name: project_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/v1.ProjectConfigResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取项目配置
      tags:
      - Integration Admin
    put:
      consumes:
      - application/json
      parameters:
      - description: Bearer Admin JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: 项目 ID
        in: path
        name: project_id
        required: true
        type: string
      - description: 项目基本信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateProjectReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新项目
      tags:
      - Integration Admin
  /api/v1/integrations/projects/{project_id}/config:
    put:
      consumes:
      - application/json
      description: 一次性替换项目基本信息、公钥、飞书表配置和表级 Scope。未提交的旧公钥或表配置会被软删除。
      parameters:
      - description: Bearer Admin JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: 项目 ID
        in: path
        name: project_id
        required: true
        type: string
      - description: 完整项目配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.UpdateProjectConfigReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: 全量更新项目配置
      tags:
      - Integration Admin
...
  /api/v1/message/trigger:
    post:
      consumes:
      - application/json
      description: 根据项目和表格标识，将表格写入通知通道以触发下游消费
      operationId: trigger-notification
      parameters:
      - description: 触发通知请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.TriggerNotificationReq'
      produces:
      - application/json
      responses:
        "200":
          description: 触发成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      summary: 手动触发通知
      tags:
      - Message
...
swagger: "2.0"